added PathsService, that will manage every process that has to do with the folders and resources. Added methods to do that

+++ src/main/java/xyz/drena/exports/ExportToImage.java
@@ -16,8 +16,8 @@ public class ExportToImage implements Exportable {
     public void export(LinkedList<ExportUnits> exportUnits, String fileName) {
 
         BufferedImage bufferedImage = new BufferedImage(
-                Constants.GENERATOR_LAB_COLUMNS*Constants.EXPORT_VIEW_UNIT_SIZE,
-                Constants.GENERATOR_LAB_ROWS*Constants.EXPORT_VIEW_UNIT_SIZE,
+                (int) Constants.GENERATOR_LAB_COLUMNS*Constants.EXPORT_VIEW_UNIT_SIZE,
+                (int) Constants.GENERATOR_LAB_ROWS*Constants.EXPORT_VIEW_UNIT_SIZE,
                 BufferedImage.TYPE_INT_RGB
         );
